Add commonly used test models as fixtures
At first, we avoided using fixtures in order to keep specs
self-contained and avoid some of the increased cognitive load that can
come with their use. However, we are re-using the same models enough
that it makes sense to DRY it up at this point. In this commit, we've
added fixtures for rats, users, and their association model, rat_users.
We are not adding photo-related fixtures for now.